requirements.txt
matplotlib==3.4.2
tensorflow==2.5.0
imageio==2.9.0
Pillow==8.3.0
